Refactoring output for GET all bucketlists to use marshmallow schema
