+++ bucketlist/app/models.py
@@ -1,5 +1,4 @@
 # This is where you define the models of your application.
-# This may be split into several modules in the same way as views.py
 
 # datetime.fromtimestamp(time.time())
