+++ requirements.txt
@@ -15,6 +15,7 @@ flask-heroku==0.1.9
 Flask-HTTPAuth==3.2.2
 Flask-Login==0.4.0
 Flask-Mail==0.9.1
+flask-marshmallow==0.7.0
 Flask-Migrate==2.0.3
 Flask-Principal==0.4.0
 Flask-RESTful==0.3.5
@@ -22,10 +23,15 @@ Flask-Script==2.0.5
 Flask-Security==1.7.5
 Flask-SocketIO==2.8.6
 Flask-SQLAlchemy==2.2
+Flask-WTF==0.14.2
 idna==2.5
 itsdangerous==0.24
+Jinja2==2.9.6
+jsonpickle==0.9.4
 Mako==1.0.6
 MarkupSafe==1.0
+marshmallow==2.13.5
+marshmallow-sqlalchemy==0.13.1
 migrate==0.3.8
 nose==1.3.7
 packaging==16.8
@@ -50,4 +56,5 @@ sqlalchemy-migrate==0.11.0
 sqlparse==0.2.3
 Tempita==0.5.2
 termcolor==1.1.0
-Werkzeug==0.12.1
+Werkzeug==0.12.2
+WTForms==2.1
